+++ src/WeightedDecisionMakerTest.java
@@ -13,28 +13,28 @@ public class WeightedDecisionMakerTest {
 
 	@Test
 	public void dummy() {
-		decision = dm.decide(new double[]{1, 1, 1});
+		decision = dm.decide(new int[]{1, 1, 1});
 		assertEquals(1, decision, 0.0);
 	}
 
 	@Test
 	public void simpleMajority() {
-		decision = dm.decide(new double[]{1, 0, 1});
+		decision = dm.decide(new int[]{1, 0, 1});
 		assertEquals(1, decision, 0.0);
 
-		decision = dm.decide(new double[]{1, 0, 0});
+		decision = dm.decide(new int[]{1, 0, 0});
 		assertEquals(0, decision, 0.0);
 
-		decision = dm.decide(new double[]{0, 1, 1});
+		decision = dm.decide(new int[]{0, 1, 1});
 		assertEquals(1, decision, 0.0);
 
-		decision = dm.decide(new double[]{1, 0, 0});
+		decision = dm.decide(new int[]{1, 0, 0});
 		assertEquals(0, decision, 0.0);
 	}
 
 	@Test
 	public void draw() {
-		decision = dm.decide(new double[]{1, 1, 0});
+		decision = dm.decide(new int[]{1, 1, 0});
 		assertEquals(1, decision, 0.0);
 	}
 }
